refactor(settings): extract HelpSection with onboarding + update buttons
Create HelpSection component combining restart onboarding guide and
check-for-updates (Electron only) buttons. Replace inline Help section
in AdvancedSettings and UpdateSection in SimpleSettings with HelpSection.
